Accurately compute reflections
This finally resolves the disparity between HDR and LDR maps correctly, with a simple solution. We hijack the last render's c30 constant to obtain the cLightScale rendering parameter in order to extract the proper environment map scaling factor for the map.

Second, we mark the cubemap texture as sRGB such that it is properly decoded to linear RGB space so that it does not contaminate our linear RGB shading.

D3D9 takes care of the output from there, and now our reflections are 1:1 with GMod!